+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project are documented here. Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/); this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Added Manhattan, correlation, Canberra, and Bray-Curtis distance metrics.
+
 ## [0.4.1] - 2026-08-17
 
 ### Fixed